+++ examples/wrapper-components/vue-javascript/src/components/Alert.vue
