+++ examples/wrapper-components/vue-javascript/src/components/Checkbox.vue
