+++ examples/wrapper-components/vue-javascript/src/components/ProgressBar.vue
