+++ examples/wrapper-components/vue-javascript/src/components/Search.vue
