+++ examples/wrapper-components/vue-typescript/src/App.vue
@@ -12,12 +12,12 @@
 
 
 <script lang="ts" setup>
-import Alert from './wrapper-components/Alert.vue'
-import Button from './wrapper-components/Button.vue'
-import Checkbox from './wrapper-components/Checkbox.vue'
-import ProgressBar from './wrapper-components/ProgressBar.vue'
-import Search from './wrapper-components/Search.vue'
-import TextInput from './wrapper-components/TextInput.vue'
+import Alert from './components/Alert.vue'
+import Button from './components/Button.vue'
+import Checkbox from './components/Checkbox.vue'
+import ProgressBar from './components/ProgressBar.vue'
+import Search from './components/Search.vue'
+import TextInput from './components/TextInput.vue'
 
 
 </script>
