+++ examples/wrapper-components/vue-typescript/src/components/Alert.vue
