+++ examples/wrapper-components/vue-typescript/src/components/Checkbox.vue
