+++ examples/wrapper-components/vue-typescript/src/components/ProgressBar.vue
