+++ examples/wrapper-components/vue-typescript/src/components/Search.vue
